% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Null_models_fcts.R
\name{compute.null.model.FD}
\alias{compute.null.model.FD}
\title{Compute null models for FD metrics}
\usage{
compute.null.model.FD(sp_faxes_coord, sp_asb_df, nb_asb_rep)
}
\arguments{
\item{sp_faxes_coord}{the array with species coordinates of functional axes -
with only the PC wanted. Saved in the \code{2_a_Compute_..._TD_FD_PD.R}.}

\item{sp_asb_df}{a data frame with species in columns and asb (grid cells)
in rows.}

\item{nb_asb_rep}{a number with the number of null assemblages to compute for
each grid cell.}
}
\value{
a list of data frames - as many as FD metrics - with the grid cells
in rows and the values of the metric for each null asb in columns.
}
\description{
Compute null models for FD metrics
}
